+++ api/src/main/java/org/pmiops/workbench/api/ConceptSetsController.java
@@ -22,7 +22,7 @@ import org.pmiops.workbench.conceptset.ConceptSetMapper;
 import org.pmiops.workbench.conceptset.ConceptSetService;
 import org.pmiops.workbench.db.dao.ConceptSetDao;
 import org.pmiops.workbench.db.dao.UserRecentResourceService;
-import org.pmiops.workbench.db.model.CommonStorageEnums;
+import org.pmiops.workbench.db.model.DbStorageEnums;
 import org.pmiops.workbench.db.model.DbConceptSet;
 import org.pmiops.workbench.db.model.DbUser;
 import org.pmiops.workbench.db.model.DbWorkspace;
@@ -158,7 +158,7 @@ public class ConceptSetsController implements ConceptSetsApiDelegate {
         workspaceService.getWorkspaceEnforceAccessLevelAndSetCdrVersion(
             workspaceNamespace, workspaceId, WorkspaceAccessLevel.READER);
     short surveyId =
-        CommonStorageEnums.surveysToStorage(Surveys.fromValue(surveyName.toUpperCase()));
+        DbStorageEnums.surveysToStorage(Surveys.fromValue(surveyName.toUpperCase()));
     List<DbConceptSet> conceptSets =
         conceptSetService.findByWorkspaceIdAndSurvey(workspace.getWorkspaceId(), surveyId);
     ConceptSetListResponse response = new ConceptSetListResponse();
@@ -213,7 +213,7 @@ public class ConceptSetsController implements ConceptSetsApiDelegate {
                   Domain domain =
                       Domain.PHYSICALMEASUREMENT.equals(domainEnum)
                           ? Domain.PHYSICALMEASUREMENT
-                          : CommonStorageEnums.domainIdToDomain(concept.getDomainId());
+                          : DbStorageEnums.domainIdToDomain(concept.getDomainId());
                   return !domainEnum.equals(domain);
                 })
             .collect(Collectors.toList());
