+++ api/src/main/java/org/pmiops/workbench/db/dao/DataSetServiceImpl.java
@@ -30,7 +30,7 @@ import org.pmiops.workbench.cohortbuilder.CohortQueryBuilder;
 import org.pmiops.workbench.cohortbuilder.ParticipantCriteria;
 import org.pmiops.workbench.config.CdrBigQuerySchemaConfig;
 import org.pmiops.workbench.config.CdrBigQuerySchemaConfigService;
-import org.pmiops.workbench.db.model.CommonStorageEnums;
+import org.pmiops.workbench.db.model.DbStorageEnums;
 import org.pmiops.workbench.db.model.DbCohort;
 import org.pmiops.workbench.db.model.DbConceptSet;
 import org.pmiops.workbench.db.model.DbDataset;
@@ -802,7 +802,7 @@ public class DataSetServiceImpl implements DataSetService, GaugeDataCollector {
         ImmutableList.copyOf(conceptBigQueryService.getSurveyQuestionConceptIds());
     final DbConceptSet surveyConceptSet = new DbConceptSet();
     surveyConceptSet.setName("All Surveys");
-    surveyConceptSet.setDomain(CommonStorageEnums.domainToStorage(Domain.SURVEY));
+    surveyConceptSet.setDomain(DbStorageEnums.domainToStorage(Domain.SURVEY));
     surveyConceptSet.setConceptIds(ImmutableSet.copyOf(conceptIds));
     return surveyConceptSet;
   }
