+++ api/src/main/java/org/pmiops/workbench/db/model/DbDataset.java
@@ -205,10 +205,10 @@ public class DbDataset {
 
   @Transient
   public PrePackagedConceptSetEnum getPrePackagedConceptSetEnum() {
-    return CommonStorageEnums.prePackageConceptSetsFromStorage(prePackagedConceptSet);
+    return DbStorageEnums.prePackageConceptSetsFromStorage(prePackagedConceptSet);
   }
 
   public void setPrePackagedConceptSetEnum(PrePackagedConceptSetEnum domain) {
-    this.prePackagedConceptSet = CommonStorageEnums.prePackageConceptSetsToStorage(domain);
+    this.prePackagedConceptSet = DbStorageEnums.prePackageConceptSetsToStorage(domain);
   }
 }
