+++ api/src/main/java/org/pmiops/workbench/utils/mappers/CohortMapper.java
@@ -3,7 +3,7 @@ package org.pmiops.workbench.utils.mappers;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.pmiops.workbench.db.dao.UserDao;
-import org.pmiops.workbench.db.model.CommonStorageEnums;
+import org.pmiops.workbench.db.model.DbStorageEnums;
 import org.pmiops.workbench.db.model.DbCohort;
 import org.pmiops.workbench.model.Cohort;
 import org.pmiops.workbench.model.DataAccessLevel;
@@ -24,6 +24,6 @@ public interface CohortMapper {
   Cohort dbModelToClient(DbCohort destination);
 
   public static DataAccessLevel storageEnumToDataAccessLevel(Short ordinal) {
-    return CommonStorageEnums.dataAccessLevelFromStorage(ordinal);
+    return DbStorageEnums.dataAccessLevelFromStorage(ordinal);
   }
 }
